prices changed to long, hottest returns list

+++ src/main/java/com/rahnemacollege/domain/AddAuctionDomain.java
@@ -8,16 +8,16 @@ import lombok.Data;
 public class AddAuctionDomain {
     private String title;
     private String description;
-    private int basePrice;
+    private long basePrice;
     private long date;
     private int categoryId;
     private int maxNumber;
 
-    public AddAuctionDomain(Auction auction){
+    public AddAuctionDomain(Auction auction) {
         title = auction.getTitle();
         description = auction.getDescription();
         basePrice = auction.getBasePrice();
-         date = auction.getDate().getTime();
+        date = auction.getDate().getTime();
         categoryId = auction.getCategory().getId();
         maxNumber = auction.getMaxNumber();
     }
